Created items table

+++ itemsDB.sql
@@ -1,5 +1,5 @@
 CREATE TABLE items (
 	item VARCHAR(20) NOT NULL,
-    price INT(10) NOT NULL,
-    quantity INT(20) NOT NULL
+    price DECIMAL(4, 2) NOT NULL,
+    quantity INT NOT NULL
 );
